Add database_schema.sql and some validations to crud of user

+++ src/main/java/com/stocks/project/repository/UserRepository.java
@@ -1,5 +1,7 @@
 package com.stocks.project.repository;
 
+import com.stocks.project.exception.NoFirstNameException;
+import com.stocks.project.exception.NoSuchUserException;
 import com.stocks.project.model.User;
 import com.stocks.project.utils.UserMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,6 +15,7 @@ import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Repository
 public class UserRepository {
@@ -28,10 +31,7 @@ public class UserRepository {
     public List<User> findAll() {
         List<User> users = new ArrayList<>();
         String query = "SELECT * FROM stocks_user;";
-        try(
-                Connection connection = dataSource.getConnection();
-                PreparedStatement preparedStatement = connection.prepareStatement(query)
-        ) {
+        try (Connection connection = dataSource.getConnection(); PreparedStatement preparedStatement = connection.prepareStatement(query)) {
             ResultSet resultSet = preparedStatement.executeQuery();
             while (resultSet.next()) {
                 users.add(userMapper.mapRow(resultSet));
@@ -42,13 +42,10 @@ public class UserRepository {
         return users;
     }
 
-    public User findById(int id) {
+    public Optional<User> findById(int id) {
         User user = null;
         String query = "SELECT * FROM stocks_user WHERE user_id = ?;";
-        try(
-                Connection connection = dataSource.getConnection();
-                PreparedStatement preparedStatement = connection.prepareStatement(query)
-        ) {
+        try (Connection connection = dataSource.getConnection(); PreparedStatement preparedStatement = connection.prepareStatement(query)) {
             preparedStatement.setInt(1, id);
             ResultSet resultSet = preparedStatement.executeQuery();
             if (resultSet.next()) {
@@ -57,17 +54,16 @@ public class UserRepository {
         } catch (SQLException e) {
             e.printStackTrace();
         }
-        return user;
+        return Optional.ofNullable(user);
     }
 
-    public User createUser(User newUser) {
-        User user = null;
+    public Optional<User> createUser(User newUser) throws NoFirstNameException {
+        if (newUser.getFirstName() == null) {
+            throw new NoFirstNameException("Column 'first_name' is required to create a user.");
+        }
+        Optional<User> user = Optional.empty();
         String query = "INSERT INTO stocks_user (first_name, second_name, birthday) VALUES (?, ?, ?);";
-        try(
-                Connection connection = dataSource.getConnection();
-                PreparedStatement preparedStatement =
-                        connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS)
-        ) {
+        try (Connection connection = dataSource.getConnection(); PreparedStatement preparedStatement = connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS)) {
             preparedStatement.setString(1, newUser.getFirstName());
             preparedStatement.setString(2, newUser.getSecondName());
             preparedStatement.setDate(3, newUser.getBirthday());
@@ -83,16 +79,13 @@ public class UserRepository {
         return user;
     }
 
-    public void delete(int userId) {
+    public void delete(int userId) throws NoSuchUserException {
+        if (findById(userId).isEmpty()) {
+            throw new NoSuchUserException();
+        }
         String query = "DELETE FROM stocks_user WHERE user_id = ?;";
         String queryToDeleteInfo = "DELETE FROM security_info WHERE user_id = ?;";
-        try(
-                Connection connection = dataSource.getConnection();
-                PreparedStatement deleteUserPrepStatement =
-                        connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS);
-                PreparedStatement deleteSecurityInfoPrepStatement =
-                        connection.prepareStatement(queryToDeleteInfo)
-        ) {
+        try (Connection connection = dataSource.getConnection(); PreparedStatement deleteUserPrepStatement = connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS); PreparedStatement deleteSecurityInfoPrepStatement = connection.prepareStatement(queryToDeleteInfo)) {
             try {
                 connection.setAutoCommit(false);
                 deleteUserPrepStatement.setInt(1, userId);
@@ -110,14 +103,24 @@ public class UserRepository {
         }
     }
 
-    public User updateUser(User updatedUser, int userId) {
-        User user = null;
+    public Optional<User> updateUser(User updatedUser, int userId) throws NoSuchUserException {
+        if (findById(userId).isEmpty()) {
+            throw new NoSuchUserException();
+        }
+        User oldUser = findById(userId).get();
+        if (updatedUser.getFirstName() == null) {
+            updatedUser.setFirstName(oldUser.getFirstName());
+        }
+        if (updatedUser.getSecondName() == null) {
+            updatedUser.setSecondName(oldUser.getSecondName());
+        }
+        if (updatedUser.getBirthday() == null) {
+            updatedUser.setBirthday(oldUser.getBirthday());
+        }
+
+        Optional<User> user = Optional.empty();
         String query = "UPDATE stocks_user SET first_name = ?, second_name = ?, birthday = ? WHERE user_id = ?;";
-        try(
-                Connection connection = dataSource.getConnection();
-                PreparedStatement preparedStatement =
-                        connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS)
-        ) {
+        try (Connection connection = dataSource.getConnection(); PreparedStatement preparedStatement = connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS)) {
             preparedStatement.setString(1, updatedUser.getFirstName());
             preparedStatement.setString(2, updatedUser.getSecondName());
             preparedStatement.setDate(3, updatedUser.getBirthday());
